+++ src/test/java/com/hoangluongtran0309/releaseflow/release/ReleaseNoteDigestTest.java
@@ -215,6 +215,7 @@ class ReleaseNoteDigestTest {
                 null,
                 narratives,
                 null,
+                null,
                 null
         );
     }