+++ src/test/java/com/hoangluongtran0309/releaseflow/release/ReleaseNotePreviewTest.java
@@ -107,6 +107,7 @@ class ReleaseNotePreviewTest {
                 null,
                 Map.of(),
                 null,
+                null,
                 null
         );
     }